Updated makefile: 1. build libraries for statusbar as well; 2. make install
Now the makefile also builds shared and static libraries for statusbar
in addition to progressbar.

I also added a `make install` option that copies the libraries to system-wide
directories. In this way, the user can easily incorporate progressbar in
their C code without having to include the source files in their code
directory.